producer/requirements.txt
python-dotenv
faker
jsonschema
confluent_kafka
yfinance